refactor(quiz): R15 useQuizAnalytics + R16 useContentTree extractions
R15: Extract useQuizAnalytics hook from QuizAnalyticsPanel.tsx
- Data loading (questions + attempts in parallel)
- 4 computed analytics (diffData, typeData, questionStats, globalStats)
- Color maps moved to hook module
- QuizAnalyticsPanel.tsx reduced to pure UI (~170 lines)

R16: Extract useContentTree hook from useQuizCascade.tsx
- Content tree types (Course/Semester/Section/Topic)
- Membership-filtered tree loading
- Reusable for flashcards, summaries, any cascade selector
- useQuizCascade.tsx reduced to ~200 lines of selection logic

No functionality changes. All exports preserved for backward compat.
